backend/my_project/pyproject.toml
[project]
name = "my-project"
version = "0.1.0"
description = "The Rental Farm — Tree rental ecommerce backend"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.135.2",
    "uvicorn[standard]>=0.42.0",
    "sqlalchemy>=2.0",
    "psycopg2-binary>=2.9",
    "python-dotenv>=1.0",
    "alembic>=1.13",
    "pydantic[email]>=2.0",
    "PyJWT>=2.8",
    "bcrypt>=4.0",
    "razorpay>=1.4.2",
    "stripe>=7.9.0",
]
